API_request_automate_msg_daily.py
create an account in PythonAnywhere to perform this program. This program is to automate the message in our daily day to day life daily @same time each day we can get weather update message in your location based on what longitude and latitude given by you once we have run set this task of this program but read the instruction to perform this code work.